requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
grpcio==1.66.2
h5py==3.12.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.1.2
pandas==2.2.3
pandas-stubs==2.2.3.241009
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.20.0
sift_stack_py==0.2.2
six==1.16.0
sqlparse==0.5.1
types-protobuf==5.28.0.20240924
types-pytz==2024.2.0.20241003
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
